backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.8.0
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.12
resend==2.1.0

